==> 03_SpamMessages/01_Load_and_Preprocess.ipynb <==
# ! python -m spacy download en_core_web_lg

from spam import *

msg_df = import_emails(root_dir="./SpamAssassinMessages", 
                       encoding="cp437", 
                       positive_indicator="spam")

msg_df.head(2)

preprocessed_df = text_preprocessing(df=msg_df)

preprocessed_df.loc[:, [c for c in preprocessed_df.columns if "text" in c]].head()

preprocessed_df.to_csv("./initial_preprocess_emails_092722_full.csv", index=False)

from sklearn.model_selection import train_test_split
train_df, test_df = train_test_split(preprocessed_df, 
                                     test_size=0.075, 
                                     stratify=preprocessed_df['label'].to_numpy(), 
                                     random_state=77)

train_df.to_csv("./initial_preprocess_emails_092722_train.csv", index=False)
test_df.to_csv("./initial_preprocess_emails_092722_test.csv", index=False)

preprocessed_df.shape
