==> 03_SpamMessages/02_EDA.ipynb <==
from spam import *
import plotly.express as px

df = pd.read_csv("./initial_preprocess_emails_092722_full.csv")

df.head(2)

f_dists = get_all_freq_dicts(df=df, 
                             text_columns=[col for col in df.columns if col.startswith("text")])
print(f_dists.keys())

f_dists['text']['label_0'].most_common(5)

# fd = load_gs_from_pickle(pickle_filepath="./word_counts/word_counts_20220926_1505.pkl")
# fd['text']['label_0'].most_common(5)
