==> 03_SpamMessages/03_Baseline_Models.ipynb <==
from spam import *

# df = pd.read_csv("./initial_preprocess_emails_092722_full.csv")
df = pd.read_csv("./initial_preprocess_emails_092722_train.csv")

# text_columns = [col for col in df.columns if col.startswith("text")]

# pipe = Pipeline([("tfidf", TfidfVectorizer()),
#                  ("nb", MultinomialNB())])

# base_results = get_baseline_model_results(df=df, 
#                                           text_columns=text_columns, 
#                                           target_column="label",
#                                           n_jobs=30,
#                                           estimator=pipe, 
#                                           metrics=["f1", "accuracy", "precision", "recall"])

# PATH = "./models/" + get_gs_save_name("baseline_models")
# with open(PATH, 'wb') as f:
#     pickle.dump(base_results, f)

with open("./models/baseline_models_20220927_0837.pkl", 'rb') as f:
    base_results = pickle.load(f)

base_f1_df = create_baseline_performance_df(baseline_results=base_results, metric="test_f1")
base_f1_df

base_acc_df = create_baseline_performance_df(baseline_results=base_results, metric="test_accuracy")
base_acc_df

best_preproces = 'text_lem_spd'

pipe = Pipeline([("tfidf", TfidfVectorizer()),
                 ("nb", MultinomialNB())])

X = df.loc[:, best_preproces].copy(deep=True).astype(str).to_numpy().ravel()
y = df["label"].to_numpy()

pred_df = error_analysis(model=pipe.fit(X,y), 
                         X=X, 
                         y=y, 
                         dataset_type="Full")

# g = run_gridsearch(X, y, 
#                    estimator, 
#                    param_grid, 
#                    n_jobs=30,
#                    scoring=["f1", "accuracy", "precision", "recall"], 
#                    save_name=get_gs_save_name(model_name=f"nb"))
