==> 03_SpamMessages/04_Naive_Bayes_GS.ipynb <==
from spam import *

df = pd.read_csv("./initial_preprocess_emails_092722_train.csv")

# with open("./word_counts/word_counts/word_counts_20220927_0031.pkl", 'rb') as f:
#     p = pickle.load(f)
#len(p['text_clean_lp']['all'].keys())

# pipe = Pipeline([("tfidf", TfidfVectorizer()),
#                  ("nb", MultinomialNB())])
# target = "text_clean_lp2c_lem_sa"
# pipe.fit(X=df[target].astype(str).to_numpy().ravel(), y=df["label"].to_numpy())

#### all stop

features = "text_clean_lp2c_lem_sa"

# pipe = Pipeline([("tfidf", TfidfVectorizer()),
#                  ("nb", MultinomialNB())])

# parameter_grid = [{'tfidf__analyzer' : ['word'],
#                   'tfidf__ngram_range' : [(1,1), (1,2), (1,3)],
#                   'tfidf__max_df':[0.8, 1.0],
#                   'tfidf__max_features':[125_000, 75_000, 10_000, None],
#                   'nb__alpha' : np.logspace(-2,0,20), 
#                   'nb__fit_prior' : [True], 
#                   'nb__class_prior':[None, [0.4,0.6], [0.6,0.4]]}, 
#                   {'tfidf__analyzer' : ['word'],
#                   'tfidf__ngram_range' : [(1,1), (1,2), (1,3)],
#                   'tfidf__max_df':[0.8, 1.0],
#                   'tfidf__max_features':[125_000, 75_000, 10_000, None],
#                   'nb__alpha' : np.logspace(-2,0,20), 
#                   'nb__fit_prior' : [False], 
#                   'nb__class_prior':[None]}]

# g_sa = run_gridsearch(X=df[features].astype(str).to_numpy().ravel(), 
#                    y=df['label'].to_numpy(), 
#                    folds=10,
#                    estimator=pipe, 
#                    param_grid=parameter_grid, 
#                    n_jobs=40,
#                    scoring=["f1", "accuracy", "precision", "recall"], 
#                    save_name=get_gs_save_name(model_name=f"nb_{features}"))

PATH=f"./models/nb_text_clean_lp2c_lem_sa_20220927_1220.pkl"
g_sa = load_gs_from_pickle(pickle_filepath=PATH)
g_sa_df = gs_to_clean_df(g_sa.cv_results_, sort_metric="mean_test_f1")
g_sa_df.loc[:,[c for c in g_sa_df.columns if "std" not in c]].head()

#### spacy default stop

features="text_clean_lp2c_lem_ssd"

# pipe2 = Pipeline([("tfidf", TfidfVectorizer()),
#                  ("nb", MultinomialNB())])

# parameter_grid = [{'tfidf__analyzer' : ['word'],
#                   'tfidf__ngram_range' : [(1,1), (1,2), (1,3)],
#                   'tfidf__max_df':[0.8, 1.0],
#                   'tfidf__max_features':[125_000, 75_000, 10_000, None],
#                   'nb__alpha' : np.logspace(-2,0,20), 
#                   'nb__fit_prior' : [True], 
#                   'nb__class_prior':[None, [0.4,0.6], [0.6,0.4]]}, 
#                   {'tfidf__analyzer' : ['word'],
#                   'tfidf__ngram_range' : [(1,1), (1,2), (1,3)],
#                   'tfidf__max_df':[0.8, 1.0],
#                   'tfidf__max_features':[125_000, 75_000, 10_000, None],
#                   'nb__alpha' : np.logspace(-2,0,20), 
#                   'nb__fit_prior' : [False], 
#                   'nb__class_prior':[None]}]

# g_ssd = run_gridsearch(X=df[features].astype(str).to_numpy().ravel(), 
#                    y=df['label'].to_numpy(), 
#                    folds=10,
#                    estimator=pipe2, 
#                    param_grid=parameter_grid, 
#                    n_jobs=40,
#                    scoring=["f1", "accuracy", "precision", "recall"],                    
#                    save_name=get_gs_save_name(model_name=f"nb_{features}"))

PATH=f"./models/nb_text_clean_lp2c_lem_snltk_20220927_1446.pkl"
g_ssd = load_gs_from_pickle(pickle_filepath=PATH)
g_ssd_df = gs_to_clean_df(g_ssd.cv_results_, sort_metric="mean_test_f1")
g_ssd_df.loc[:,[c for c in g_ssd_df.columns if "std" not in c]].head()

#### nltk stop

features="text_clean_lp2c_lem_snltk"

# pipe3 = Pipeline([("tfidf", TfidfVectorizer()),
#                  ("nb", MultinomialNB())])

# parameter_grid = [{'tfidf__analyzer' : ['word'],
#                   'tfidf__ngram_range' : [(1,1), (1,2), (1,3)],
#                   'tfidf__max_df':[0.8, 1.0],
#                   'tfidf__max_features':[125_000, 75_000, 10_000, None],
#                   'nb__alpha' : np.logspace(-2,0,20), 
#                   'nb__fit_prior' : [True], 
#                   'nb__class_prior':[None, [0.4,0.6], [0.6,0.4]]}, 
#                   {'tfidf__analyzer' : ['word'],
#                   'tfidf__ngram_range' : [(1,1), (1,2), (1,3)],
#                   'tfidf__max_df':[0.8, 1.0],
#                   'tfidf__max_features':[125_000, 75_000, 10_000, None],
#                   'nb__alpha' : np.logspace(-2,0,20), 
#                   'nb__fit_prior' : [False], 
#                   'nb__class_prior':[None]}]

# g_nltk = run_gridsearch(X=df[features].astype(str).to_numpy().ravel(), 
#                    y=df['label'].to_numpy(), 
#                    folds=10,
#                    estimator=pipe3, 
#                    param_grid=parameter_grid, 
#                    n_jobs=40,
#                    scoring=["f1", "accuracy", "precision", "recall"],                    
#                    save_name=get_gs_save_name(model_name=f"nb_{features}"))

PATH=f"./models/nb_text_clean_lp2c_lem_ssd_20220927_1333.pkl"
g_nltk = load_gs_from_pickle(pickle_filepath=PATH)
g_nltk_df = gs_to_clean_df(g_nltk.cv_results_, sort_metric="mean_test_f1")
g_nltk_df.loc[:,[c for c in g_nltk_df.columns if "std" not in c]].head()

#### no preprocessing prior to tfidf vectorizer

features="text"

# pipe4 = Pipeline([("tfidf", TfidfVectorizer()),
#                  ("nb", MultinomialNB())])

# parameter_grid = [{'tfidf__analyzer' : ['word'],
#                   'tfidf__ngram_range' : [(1,1), (1,2), (1,3)],
#                   'tfidf__max_df':[0.8, 1.0],
#                   'tfidf__max_features':[125_000, 75_000, 10_000, None],
#                   'nb__alpha' : np.logspace(-2,0,20), 
#                   'nb__fit_prior' : [True], 
#                   'nb__class_prior':[None, [0.4,0.6], [0.6,0.4]]}, 
#                   {'tfidf__analyzer' : ['word'],
#                   'tfidf__ngram_range' : [(1,1), (1,2), (1,3)],
#                   'tfidf__max_df':[0.8, 1.0],
#                   'tfidf__max_features':[125_000, 75_000, 10_000, None],
#                   'nb__alpha' : np.logspace(-2,0,20), 
#                   'nb__fit_prior' : [False], 
#                   'nb__class_prior':[None]}]

# g_txt = run_gridsearch(X=df[features].astype(str).to_numpy().ravel(), 
#                    y=df['label'].to_numpy(), 
#                    folds=10,
#                    estimator=pipe4, 
#                    param_grid=parameter_grid, 
#                    n_jobs=40,
#                    scoring=["f1", "accuracy", "precision", "recall"],                    
#                    save_name=get_gs_save_name(model_name=f"nb_{features}"))

PATH=f"./models/nb_text_20220927_1603.pkl"
g_txt = load_gs_from_pickle(pickle_filepath=PATH)
g_txt_df = gs_to_clean_df(g_txt.cv_results_, sort_metric="mean_test_f1")
g_txt_df.loc[:,[c for c in g_txt_df.columns if "std" not in c]].head()
